+++ vite.config.ts
@@ -1,3 +1,5 @@
+import path from 'path'; // 추가
+
 import react from '@vitejs/plugin-react-swc';
 import { defineConfig } from 'vite';
 import { defineConfig as defineTestConfig, mergeConfig } from 'vitest/config';
@@ -11,6 +13,11 @@ export default mergeConfig(
           target: 'http://localhost:3000',
           changeOrigin: true,
         },
+      },
+    },
+    resolve: {
+      alias: {
+        '@': path.resolve(__dirname, './src'), // 추가
       },
     },
   }),
